+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # groupdata2 1.3.0.9000
 
+* Breaking: In `fold()`, the `k` argument can now be a multi-element vector with one `k` (number of folds) per fold column. This functionality required a minor rewrite, why you might see interchanged fold column names in comparison to the previous versions.
+
 # groupdata2 1.3.0
 
 * Breaking: The following functions now work with *grouped* `data.frames` (meaning that they are applied group-wise): `fold()`, `partition()`, `group()`, `group_factor()`, `splt()`, `balance()`, `upsample()`, `downsample()`, `differs_from_previous()`, and `find_missing_starts()`. A message is generated *once per session*, when the input is grouped, to help users understand why their code is breaking.